pyproject.toml
[project]
name = "clw"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "astral>=3.2",
    "numpy>=2.2.3",
    "openmeteo-requests>=1.3.0",
    "pandas>=2.2.3",
    "pex>=2.33.1",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "textual>=1.0.0",
    "textual-dev>=1.7.0",
    "textual-image>=0.7.0",
]

[project.urls]
Documentation = "https://github.com/philion/clw#readme"
Issues = "https://github.com/philion/clw/issues"
Source = "https://github.com/philion/clw"

[tool.coverage.run]
source_pkgs = ["clw", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
clw = ["src/clw", "*/clw/src/clw"]
tests = ["tests", "*/clw/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch]

[tool.hatch.build.targets.wheel]
packages = ["src/clw"]

[tool.hatch.version]
path = "src/clw/__about__.py"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
clw = "clw.app:main"
